✅ Staging validation passed - all tests successful
✅ Payment enhancement deployed and validated on staging
✅ Frontend pricing consistency improved
✅ Backward compatible with existing payment flow

Enhancement details:
- Frontend sends exact price user clicked to backend
- Backend uses frontend price with geo service fallback
- Added comprehensive logging and error handling
- Maintains all existing functionality

Staging test results:
- Health check: ✅ Environment: staging
- Pricing display: ✅ All 5 buttons correct (.49, .99, .99, .99, .99)
- API endpoints: ✅ Region detection and pricing working
- Price consistency: ✅ Frontend matches API exactly

+++ v4-clean/app/services/geo.py
@@ -232,9 +232,9 @@ class GeoService:
         # Most currencies use 2 decimal places (multiply by 100)
         # Some currencies like Japanese Yen don't use decimals
         if currency.upper() in ["JPY", "KRW"]:  # Zero-decimal currencies
-            return base_amount
+            return int(base_amount)
         else:  # Standard currencies
-            return base_amount * 100
+            return int(base_amount * 100)
 
 # Create geo data file if it doesn't exist
 def create_geo_data_file():
